#279 refactor problem into multiple files
