+++ pybop/__init__.py
@@ -26,7 +26,9 @@ script_path = path.dirname(__file__)
 #
 # Problem class
 #
-from ._problem import BaseProblem, FittingProblem, DesignProblem
+from .problems.base_problem import BaseProblem
+from .problems.fitting_problem import FittingProblem
+from .problems.design_problem import DesignProblem
 
 #
 # Cost function class
